fix bug

+++ api/src/main/java/zsc/edu/abouerp/api/controller/AdministratorController.java
@@ -295,8 +295,8 @@ public class AdministratorController {
     @PreAuthorize("hasAuthority('USER_DELETE')")
     public ResultBean delete(@PathVariable Integer id) {
         Administrator administrator = administratorService.findById(id).orElseThrow(UserNotFoundException::new);
-        if (administrator.getMd5() != null) {
-            log.info("delete personnel has md5========================= {}" ,administrator.getMd5());
+        if (administrator.getMd5() != null && !administrator.getMd5().isEmpty()) {
+            log.info("delete personnel has md5========================= {}", administrator.getMd5());
             Storage storage = storageService.findByMD5(administrator.getMd5());
             if (storage.getCount() - 1 == 0) {
                 fileStorageService.delete(storage.getMd5());
